--- django/src/rdwatch/dataclasses/__init__.py
@@ -1,7 +1,0 @@
-from .crop_parse import CropParse
-from .server_status import ServerStatus
-
-__all__ = [
-    'CropParse',
-    'ServerStatus',
-]
