--- django/src/rdwatch/dataclasses/server_status.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-from datetime import timedelta
-
-
-@dataclass
-class ServerStatus:
-    uptime: timedelta
-    hostname: str
-    ip: str
