+++ django/src/rdwatch/serializers/__init__.py
@@ -6,7 +6,6 @@ from .hyper_parameters import (
 )
 from .performer import PerformerSerializer
 from .region import RegionSerializer
-from .server_status import ServerStatusSerializer
 from .site_evaluation import SiteEvaluationListSerializer, SiteEvaluationSerializer
 from .site_image import SiteImageListSerializer, SiteImageSerializer
 from .site_observation import SiteObservationListSerializer, SiteObservationSerializer
@@ -19,7 +18,6 @@ __all__ = [
     'HyperParametersWriteSerializer',
     'PerformerSerializer',
     'RegionSerializer',
-    'ServerStatusSerializer',
     'SiteEvaluationListSerializer',
     'SiteEvaluationSerializer',
     'SiteObservationListSerializer',
